+++ packages/positronic/src/dsl/workflow-runner.ts
@@ -1,7 +1,7 @@
 import { WORKFLOW_EVENTS } from './constants';
 import { applyPatches } from './json-patch';
 import type { Adapter } from "../adapters/types";
-import type { ResourceLoader } from "../file-stores/types";
+import type { ResourceLoader } from "../resource-loaders/types";
 import type { SerializedStep, Workflow } from './workflow';
 import type { State } from './types';
 import type { PromptClient } from '../clients/types';
