+++ packages/positronic/src/index.ts
@@ -17,9 +17,9 @@ export type {
 } from "./dsl/workflow";
 export type { PromptClient, ResponseModel } from "./clients/types";
 export type { State } from "./dsl/types";
-export type { ResourceLoader } from "./file-stores/types";
+export type { ResourceLoader } from "./resource-loaders/types";
 export type { Shell } from "./shells/types";
 export { createPatch, applyPatches } from "./dsl/json-patch";
-export { LocalResourceLoader } from "./file-stores/local-file-store";
+export { LocalResourceLoader } from "./resource-loaders/local-resource-loader";
 export { SSHShell } from "./shells/ssh-shell";
 export { LocalShell } from "./shells/local-shell";
